+++ agentid/__init__.py
@@ -1,3 +1,3 @@
-"""AgentID package."""
+"""AgentPass package."""
 
 __version__ = "0.2.0"